viz: strip out old CPU CQT implementation
stub in HelloDraw to get ready for the PULSE
