+++ mutate-visualizer/src/audio/mod.rs
@@ -1,13 +1,6 @@
 // Copyright 2026 The MuTate Contributors
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-//! # Audio Nodes
+//! # Audio
 //!
-//! Audio first must be acquired, using the `crate::audio::raw::RawAudio` node.  As the graph
-//! interfaces evolve, so must these nodes and their types.
-
-pub mod colors;
-pub mod cqt;
-pub mod kweight;
-pub mod raw;
-pub mod rms;
+//! Select a device.  Set up stream from server to device.
